pyproject.toml
[project]
name = "architectllm"
version = "0.1.0"
description = "Training data generation pipeline for fine-tuning gpt-oss-20b as a system design architect expert"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0",
    "trafilatura>=1.12",
    "markdownify>=0.13",
    "anthropic>=0.40",
    "openai>=1.50",
    "datasets>=3.0",
    "pydantic>=2.5",
    "click>=8.1",
    "pyyaml>=6.0",
    "rich>=13.0",
    "tenacity>=8.2",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "tiktoken>=0.7",
    "lxml>=5.0",
    "aiohttp>=3.9",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4"]
playwright = ["crawlee[playwright]>=0.4"]

[project.scripts]
architectllm = "src.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"
